Switch to uniform weak handler symbols
Switch all *_vectors.S files to using uniformly-named weak handler
symbols.

This includes renaming "reset" handler function to "reset_handler" to
match others. Whatever vector is mentioned as "interrupt" in ST's RM0008
also gets "_irq" in the name.

This also prepares the code for switching to a single vectors.S file.

Thanks to Edoardo Galdi for showing how to do weak symbols for this!

Related to #1